Struct google_androidpublisher3::api::EditMethods

pub struct EditMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on edit resources. It is not used directly, but through the AndroidPublisher hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_androidpublisher3 as androidpublisher3;
 
use std::default::Default;
use oauth2;
use androidpublisher3::AndroidPublisher;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = AndroidPublisher::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `apks_addexternallyhosted(...)`, `apks_list(...)`, `apks_upload(...)`, `bundles_list(...)`, `bundles_upload(...)`, `commit(...)`, `delete(...)`, `deobfuscationfiles_upload(...)`, `details_get(...)`, `details_patch(...)`, `details_update(...)`, `expansionfiles_get(...)`, `expansionfiles_patch(...)`, `expansionfiles_update(...)`, `expansionfiles_upload(...)`, `get(...)`, `images_delete(...)`, `images_deleteall(...)`, `images_list(...)`, `images_upload(...)`, `insert(...)`, `listings_delete(...)`, `listings_deleteall(...)`, `listings_get(...)`, `listings_list(...)`, `listings_patch(...)`, `listings_update(...)`, `testers_get(...)`, `testers_patch(...)`, `testers_update(...)`, `tracks_get(...)`, `tracks_list(...)`, `tracks_patch(...)`, `tracks_update(...)` and `validate(...)`
// to build up your call.
let rb = hub.edits();

Implementations

impl<'a, C> EditMethods<'a, C>

pub fn apks_addexternallyhosted(
&self,
request: ApksAddExternallyHostedRequest,
package_name: &str,
edit_id: &str
) -> EditApkAddexternallyhostedCall<'a, C>

Create a builder to help you perform the following task:

Creates a new APK without uploading the APK itself to Google Play, instead hosting the APK at a specified URL. This function is only available to organizations using Managed Play whose application is configured to restrict distribution to the organizations.

Arguments

request - No description provided.

packageName - Package name of the app.

editId - Identifier of the edit.

pub fn apks_list(
&self,
package_name: &str,
edit_id: &str
) -> EditApkListCall<'a, C>

Create a builder to help you perform the following task:

Lists all current APKs of the app and edit.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

pub fn apks_upload(
&self,
package_name: &str,
edit_id: &str
) -> EditApkUploadCall<'a, C>

Create a builder to help you perform the following task:

Uploads an APK and adds to the current edit.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

pub fn bundles_list(
&self,
package_name: &str,
edit_id: &str
) -> EditBundleListCall<'a, C>

Create a builder to help you perform the following task:

Lists all current Android App Bundles of the app and edit.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

pub fn bundles_upload(
&self,
package_name: &str,
edit_id: &str
) -> EditBundleUploadCall<'a, C>

Create a builder to help you perform the following task:

Uploads a new Android App Bundle to this edit. If you are using the Google API client libraries, please increase the timeout of the http request before calling this endpoint (a timeout of 2 minutes is recommended). See Timeouts and Errors for an example in java.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

pub fn deobfuscationfiles_upload(
&self,
package_name: &str,
edit_id: &str,
apk_version_code: i32,
deobfuscation_file_type: &str
) -> EditDeobfuscationfileUploadCall<'a, C>

Create a builder to help you perform the following task:

Uploads a new deobfuscation file and attaches to the specified APK.

Arguments

packageName - Unique identifier for the Android app.

editId - Unique identifier for this edit.

apkVersionCode - The version code of the APK whose Deobfuscation File is being uploaded.

deobfuscationFileType - The type of the deobfuscation file.

pub fn details_get(
&self,
package_name: &str,
edit_id: &str
) -> EditDetailGetCall<'a, C>

Create a builder to help you perform the following task:

Gets details of an app.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

pub fn details_patch(
&self,
request: AppDetails,
package_name: &str,
edit_id: &str
) -> EditDetailPatchCall<'a, C>

Create a builder to help you perform the following task:

Patches details of an app.

Arguments

request - No description provided.

packageName - Package name of the app.

editId - Identifier of the edit.

pub fn details_update(
&self,
request: AppDetails,
package_name: &str,
edit_id: &str
) -> EditDetailUpdateCall<'a, C>

Create a builder to help you perform the following task:

Updates details of an app.

Arguments

request - No description provided.

packageName - Package name of the app.

editId - Identifier of the edit.

pub fn expansionfiles_get(
&self,
package_name: &str,
edit_id: &str,
apk_version_code: i32,
expansion_file_type: &str
) -> EditExpansionfileGetCall<'a, C>

Create a builder to help you perform the following task:

Fetches the expansion file configuration for the specified APK.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

apkVersionCode - The version code of the APK whose expansion file configuration is being read or modified.

expansionFileType - The file type of the file configuration which is being read or modified.

pub fn expansionfiles_patch(
&self,
request: ExpansionFile,
package_name: &str,
edit_id: &str,
apk_version_code: i32,
expansion_file_type: &str
) -> EditExpansionfilePatchCall<'a, C>

Create a builder to help you perform the following task:

Patches the APK's expansion file configuration to reference another APK's expansion file. To add a new expansion file use the Upload method.

Arguments

request - No description provided.

packageName - Package name of the app.

editId - Identifier of the edit.

apkVersionCode - The version code of the APK whose expansion file configuration is being read or modified.

expansionFileType - The file type of the expansion file configuration which is being updated.

pub fn expansionfiles_update(
&self,
request: ExpansionFile,
package_name: &str,
edit_id: &str,
apk_version_code: i32,
expansion_file_type: &str
) -> EditExpansionfileUpdateCall<'a, C>

Create a builder to help you perform the following task:

Updates the APK's expansion file configuration to reference another APK's expansion file. To add a new expansion file use the Upload method.

Arguments

request - No description provided.

packageName - Package name of the app.

editId - Identifier of the edit.

apkVersionCode - The version code of the APK whose expansion file configuration is being read or modified.

expansionFileType - The file type of the file configuration which is being read or modified.

pub fn expansionfiles_upload(
&self,
package_name: &str,
edit_id: &str,
apk_version_code: i32,
expansion_file_type: &str
) -> EditExpansionfileUploadCall<'a, C>

Create a builder to help you perform the following task:

Uploads a new expansion file and attaches to the specified APK.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

apkVersionCode - The version code of the APK whose expansion file configuration is being read or modified.

expansionFileType - The file type of the expansion file configuration which is being updated.

pub fn images_delete(
&self,
package_name: &str,
edit_id: &str,
language: &str,
image_type: &str,
image_id: &str
) -> EditImageDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the image (specified by id) from the edit.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

language - Language localization code (a BCP-47 language tag; for example, "de-AT" for Austrian German).

imageType - Type of the Image.

imageId - Unique identifier an image within the set of images attached to this edit.

pub fn images_deleteall(
&self,
package_name: &str,
edit_id: &str,
language: &str,
image_type: &str
) -> EditImageDeleteallCall<'a, C>

Create a builder to help you perform the following task:

Deletes all images for the specified language and image type. Returns an empty response if no images are found.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

language - Language localization code (a BCP-47 language tag; for example, "de-AT" for Austrian German). Providing a language that is not supported by the App is a no-op.

imageType - Type of the Image. Providing an image type that refers to no images is a no-op.

pub fn images_list(
&self,
package_name: &str,
edit_id: &str,
language: &str,
image_type: &str
) -> EditImageListCall<'a, C>

Create a builder to help you perform the following task:

Lists all images. The response may be empty.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

language - Language localization code (a BCP-47 language tag; for example, "de-AT" for Austrian German). There must be a store listing for the specified language.

imageType - Type of the Image. Providing an image type that refers to no images will return an empty response.

pub fn images_upload(
&self,
package_name: &str,
edit_id: &str,
language: &str,
image_type: &str
) -> EditImageUploadCall<'a, C>

Create a builder to help you perform the following task:

Uploads an image of the specified language and image type, and adds to the edit.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

language - Language localization code (a BCP-47 language tag; for example, "de-AT" for Austrian German). Providing a language that is not supported by the App is a no-op.

imageType - Type of the Image.

pub fn listings_delete(
&self,
package_name: &str,
edit_id: &str,
language: &str
) -> EditListingDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a localized store listing.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

language - Language localization code (a BCP-47 language tag; for example, "de-AT" for Austrian German).

pub fn listings_deleteall(
&self,
package_name: &str,
edit_id: &str
) -> EditListingDeleteallCall<'a, C>

Create a builder to help you perform the following task:

Deletes all store listings.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

pub fn listings_get(
&self,
package_name: &str,
edit_id: &str,
language: &str
) -> EditListingGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a localized store listing.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

language - Language localization code (a BCP-47 language tag; for example, "de-AT" for Austrian German).

pub fn listings_list(
&self,
package_name: &str,
edit_id: &str
) -> EditListingListCall<'a, C>

Create a builder to help you perform the following task:

Lists all localized store listings.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

pub fn listings_patch(
&self,
request: Listing,
package_name: &str,
edit_id: &str,
language: &str
) -> EditListingPatchCall<'a, C>

Create a builder to help you perform the following task:

Patches a localized store listing.

Arguments

request - No description provided.

packageName - Package name of the app.

editId - Identifier of the edit.

language - Language localization code (a BCP-47 language tag; for example, "de-AT" for Austrian German).

pub fn listings_update(
&self,
request: Listing,
package_name: &str,
edit_id: &str,
language: &str
) -> EditListingUpdateCall<'a, C>

Create a builder to help you perform the following task:

Creates or updates a localized store listing.

Arguments

request - No description provided.

packageName - Package name of the app.

editId - Identifier of the edit.

language - Language localization code (a BCP-47 language tag; for example, "de-AT" for Austrian German).

pub fn testers_get(
&self,
package_name: &str,
edit_id: &str,
track: &str
) -> EditTesterGetCall<'a, C>

Create a builder to help you perform the following task:

Gets testers.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

track - The track to read from.

pub fn testers_patch(
&self,
request: Testers,
package_name: &str,
edit_id: &str,
track: &str
) -> EditTesterPatchCall<'a, C>

Create a builder to help you perform the following task:

Patches testers.

Arguments

request - No description provided.

packageName - Package name of the app.

editId - Identifier of the edit.

track - The track to update.

pub fn testers_update(
&self,
request: Testers,
package_name: &str,
edit_id: &str,
track: &str
) -> EditTesterUpdateCall<'a, C>

Create a builder to help you perform the following task:

Updates testers.

Arguments

request - No description provided.

packageName - Package name of the app.

editId - Identifier of the edit.

track - The track to update.

pub fn tracks_get(
&self,
package_name: &str,
edit_id: &str,
track: &str
) -> EditTrackGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a track.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

track - Identifier of the track.

pub fn tracks_list(
&self,
package_name: &str,
edit_id: &str
) -> EditTrackListCall<'a, C>

Create a builder to help you perform the following task:

Lists all tracks.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

pub fn tracks_patch(
&self,
request: Track,
package_name: &str,
edit_id: &str,
track: &str
) -> EditTrackPatchCall<'a, C>

Create a builder to help you perform the following task:

Patches a track.

Arguments

request - No description provided.

packageName - Package name of the app.

editId - Identifier of the edit.

track - Identifier of the track.

pub fn tracks_update(
&self,
request: Track,
package_name: &str,
edit_id: &str,
track: &str
) -> EditTrackUpdateCall<'a, C>

Create a builder to help you perform the following task:

Updates a track.

Arguments

request - No description provided.

packageName - Package name of the app.

editId - Identifier of the edit.

track - Identifier of the track.

pub fn commit(&self, package_name: &str, edit_id: &str) -> EditCommitCall<'a, C>

Create a builder to help you perform the following task:

Commits an app edit.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

pub fn delete(&self, package_name: &str, edit_id: &str) -> EditDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes an app edit.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

pub fn get(&self, package_name: &str, edit_id: &str) -> EditGetCall<'a, C>

Create a builder to help you perform the following task:

Gets an app edit.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.

pub fn insert(
&self,
request: AppEdit,
package_name: &str
) -> EditInsertCall<'a, C>

Create a builder to help you perform the following task:

Creates a new edit for an app.

Arguments

request - No description provided.

packageName - Package name of the app.

pub fn validate(
&self,
package_name: &str,
edit_id: &str
) -> EditValidateCall<'a, C>

Create a builder to help you perform the following task:

Validates an app edit.

Arguments

packageName - Package name of the app.

editId - Identifier of the edit.