Module google_androidpublisher3::api

Structs

AndroidPublisher
Central instance to access all AndroidPublisher related resource activities

Apk
Information about an APK. The resource for ApksService.

ApkBinary
Represents the binary payload of an APK.

ApksAddExternallyHostedRequest
Request to create a new externally hosted APK.

ApksAddExternallyHostedResponse
Response for creating a new externally hosted APK.

ApksListResponse
Response listing all APKs.

AppDetails
The app details. The resource for DetailsService.

AppEdit
An app edit. The resource for EditsService.

Bundle
Information about a bundle. The resource for BundlesService.

BundlesListResponse
Response listing all bundles.

Comment
An entry of conversation between user and developer.

CountryTargeting
Country targeting specification.

DeobfuscationFile
Represents a deobfuscation file.

DeobfuscationFilesUploadResponse
Responses for the upload.

DeveloperComment
Developer entry from conversation between user and developer.

DeviceMetadata
Characteristics of the user's device.

DeviceSpec
The device spec used to generate a system APK.

EditApkAddexternallyhostedCall
Creates a new APK without uploading the APK itself to Google Play, instead hosting the APK at a specified URL. This function is only available to organizations using Managed Play whose application is configured to restrict distribution to the organizations.

EditApkListCall
Lists all current APKs of the app and edit.

EditApkUploadCall
Uploads an APK and adds to the current edit.

EditBundleListCall
Lists all current Android App Bundles of the app and edit.

EditBundleUploadCall
Uploads a new Android App Bundle to this edit. If you are using the Google API client libraries, please increase the timeout of the http request before calling this endpoint (a timeout of 2 minutes is recommended). See Timeouts and Errors for an example in java.

EditCommitCall
Commits an app edit.

EditDeleteCall
Deletes an app edit.

EditDeobfuscationfileUploadCall
Uploads a new deobfuscation file and attaches to the specified APK.

EditDetailGetCall
Gets details of an app.

EditDetailPatchCall
Patches details of an app.

EditDetailUpdateCall
Updates details of an app.

EditExpansionfileGetCall
Fetches the expansion file configuration for the specified APK.

EditExpansionfilePatchCall
Patches the APK's expansion file configuration to reference another APK's expansion file. To add a new expansion file use the Upload method.

EditExpansionfileUpdateCall
Updates the APK's expansion file configuration to reference another APK's expansion file. To add a new expansion file use the Upload method.

EditExpansionfileUploadCall
Uploads a new expansion file and attaches to the specified APK.

EditGetCall
Gets an app edit.

EditImageDeleteCall
Deletes the image (specified by id) from the edit.

EditImageDeleteallCall
Deletes all images for the specified language and image type. Returns an empty response if no images are found.

EditImageListCall
Lists all images. The response may be empty.

EditImageUploadCall
Uploads an image of the specified language and image type, and adds to the edit.

EditInsertCall
Creates a new edit for an app.

EditListingDeleteCall
Deletes a localized store listing.

EditListingDeleteallCall
Deletes all store listings.

EditListingGetCall
Gets a localized store listing.

EditListingListCall
Lists all localized store listings.

EditListingPatchCall
Patches a localized store listing.

EditListingUpdateCall
Creates or updates a localized store listing.

EditMethods
A builder providing access to all methods supported on edit resources. It is not used directly, but through the AndroidPublisher hub.

EditTesterGetCall
Gets testers.

EditTesterPatchCall
Patches testers.

EditTesterUpdateCall
Updates testers.

EditTrackGetCall
Gets a track.

EditTrackListCall
Lists all tracks.

EditTrackPatchCall
Patches a track.

EditTrackUpdateCall
Updates a track.

EditValidateCall
Validates an app edit.

ExpansionFile
An expansion file. The resource for ExpansionFilesService.

ExpansionFilesUploadResponse
Response for uploading an expansion file.

ExternallyHostedApk
Defines an APK available for this application that is hosted externally and not uploaded to Google Play. This function is only available to organizations using Managed Play whose application is configured to restrict distribution to the organizations.

Image
An uploaded image. The resource for ImagesService.

ImagesDeleteAllResponse
Response for deleting all images.

ImagesListResponse
Response listing all images.

ImagesUploadResponse
Response for uploading an image.

InAppProduct
An in-app product. The resource for InappproductsService.

InAppProductListing
Store listing of a single in-app product.

InappproductDeleteCall
Deletes an in-app product (i.e. a managed product or a subscriptions).

InappproductGetCall
Gets an in-app product, which can be a managed product or a subscription.

InappproductInsertCall
Creates an in-app product (i.e. a managed product or a subscriptions).

InappproductListCall
Lists all in-app products - both managed products and subscriptions.

InappproductMethods
A builder providing access to all methods supported on inappproduct resources. It is not used directly, but through the AndroidPublisher hub.

InappproductPatchCall
Patches an in-app product (i.e. a managed product or a subscriptions).

InappproductUpdateCall
Updates an in-app product (i.e. a managed product or a subscriptions).

InappproductsListResponse
Response listing all in-app products.

InternalAppSharingArtifact
An artifact resource which gets created when uploading an APK or Android App Bundle through internal app sharing.

InternalappsharingartifactMethods
A builder providing access to all methods supported on internalappsharingartifact resources. It is not used directly, but through the AndroidPublisher hub.

InternalappsharingartifactUploadapkCall
Uploads an APK to internal app sharing. If you are using the Google API client libraries, please increase the timeout of the http request before calling this endpoint (a timeout of 2 minutes is recommended). See Timeouts and Errors for an example in java.

InternalappsharingartifactUploadbundleCall
Uploads an app bundle to internal app sharing. If you are using the Google API client libraries, please increase the timeout of the http request before calling this endpoint (a timeout of 2 minutes is recommended). See Timeouts and Errors for an example in java.

IntroductoryPriceInfo
Contains the introductory price information for a subscription.

Listing
A localized store listing. The resource for ListingsService.

ListingsListResponse
Response listing all localized listings.

LocalizedText
Release notes specification, i.e. language and text.

OrderMethods
A builder providing access to all methods supported on order resources. It is not used directly, but through the AndroidPublisher hub.

OrderRefundCall
Refund a user's subscription or in-app purchase order.

PageInfo
Information about the current page. List operations that supports paging return only one "page" of results. This protocol buffer message describes the page that has been returned.

Price
Definition of a price, i.e. currency and units.

ProductPurchase
A ProductPurchase resource indicates the status of a user's inapp product purchase.

ProductPurchasesAcknowledgeRequest
Request for the product.purchases.acknowledge API.

PurchaseMethods
A builder providing access to all methods supported on purchase resources. It is not used directly, but through the AndroidPublisher hub.

PurchaseProductAcknowledgeCall
Acknowledges a purchase of an inapp item.

PurchaseProductGetCall
Checks the purchase and consumption status of an inapp item.

PurchaseSubscriptionAcknowledgeCall
Acknowledges a subscription purchase.

PurchaseSubscriptionCancelCall
Cancels a user's subscription purchase. The subscription remains valid until its expiration time.

PurchaseSubscriptionDeferCall
Defers a user's subscription purchase until a specified future expiration time.

PurchaseSubscriptionGetCall
Checks whether a user's subscription purchase is valid and returns its expiry time.

PurchaseSubscriptionRefundCall
Refunds a user's subscription purchase, but the subscription remains valid until its expiration time and it will continue to recur.

PurchaseSubscriptionRevokeCall
Refunds and immediately revokes a user's subscription purchase. Access to the subscription will be terminated immediately and it will stop recurring.

PurchaseVoidedpurchaseListCall
Lists the purchases that were canceled, refunded or charged-back.

Review
An Android app review.

ReviewGetCall
Gets a single review.

ReviewListCall
Lists all reviews.

ReviewMethods
A builder providing access to all methods supported on review resources. It is not used directly, but through the AndroidPublisher hub.

ReviewReplyCall
Replies to a single review, or updates an existing reply.

ReviewReplyResult
The result of replying/updating a reply to review.

ReviewsListResponse
Response listing reviews.

ReviewsReplyRequest
Request to reply to review or update existing reply.

ReviewsReplyResponse
Response on status of replying to a review.

SubscriptionCancelSurveyResult
Information provided by the user when they complete the subscription cancellation flow (cancellation reason survey).

SubscriptionDeferralInfo
A SubscriptionDeferralInfo contains the data needed to defer a subscription purchase to a future expiry time.

SubscriptionPriceChange
Contains the price change information for a subscription that can be used to control the user journey for the price change in the app. This can be in the form of seeking confirmation from the user or tailoring the experience for a successful conversion.

SubscriptionPurchase
A SubscriptionPurchase resource indicates the status of a user's subscription purchase.

SubscriptionPurchasesAcknowledgeRequest
Request for the purchases.subscriptions.acknowledge API.

SubscriptionPurchasesDeferRequest
Request for the purchases.subscriptions.defer API.

SubscriptionPurchasesDeferResponse
Response for the purchases.subscriptions.defer API.

SystemApksListResponse
Response to list previously created system APK variants.

SystemapkMethods
A builder providing access to all methods supported on systemapk resources. It is not used directly, but through the AndroidPublisher hub.

SystemapkVariantCreateCall
Creates an APK which is suitable for inclusion in a system image from an already uploaded Android App Bundle.

SystemapkVariantDownloadCall
Downloads a previously created system APK which is suitable for inclusion in a system image.

SystemapkVariantGetCall
Returns a previously created system APK variant.

SystemapkVariantListCall
Returns the list of previously created system APK variants.

Testers
The testers of an app. The resource for TestersService.

Timestamp
A Timestamp represents a point in time independent of any time zone or local calendar, encoded as a count of seconds and fractions of seconds at nanosecond resolution. The count is relative to an epoch at UTC midnight on January 1, 1970.

TokenPagination
Pagination information returned by a List operation when token pagination is enabled. List operations that supports paging return only one "page" of results. This protocol buffer message describes the page that has been returned. When using token pagination, clients should use the next/previous token to get another page of the result. The presence or absence of next/previous token indicates whether a next/previous page is available and provides a mean of accessing this page. ListRequest.page_token should be set to either next_page_token or previous_page_token to access another page.

Track
A track configuration. The resource for TracksService.

TrackRelease
A release within a track.

TracksListResponse
Response listing all tracks.

UserComment
User entry from conversation between user and developer.

UsesPermission
A permission used by this APK.

Variant
APK that is suitable for inclusion in a system image. The resource of SystemApksService.

VoidedPurchase
A VoidedPurchase resource indicates a purchase that was either canceled/refunded/charged-back.

VoidedPurchasesListResponse
Response for the voidedpurchases.list API.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.